+++ automotive/src/main/java/au/com/shiftyjelly/pocketcasts/AutomotiveSettingsActivity.kt
@@ -99,7 +99,10 @@ class AutomotiveSettingsActivity : AppCompatActivity(), FragmentHostListener {
     override fun closePodcastsToRoot() {
     }
 
-    override fun openPodcastPage(uuid: String) {
+    override fun openProfile() {
+    }
+
+    override fun openPodcastPage(uuid: String, sourceView: String?) {
     }
 
     override fun openCloudFiles() {