+++ modules/features/discover/src/main/java/au/com/shiftyjelly/pocketcasts/discover/viewmodel/DiscoverViewModel.kt
@@ -3,11 +3,13 @@ package au.com.shiftyjelly.pocketcasts.discover.viewmodel
 import android.content.res.Resources
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
+import androidx.lifecycle.viewModelScope
 import au.com.shiftyjelly.pocketcasts.analytics.AnalyticsEvent
 import au.com.shiftyjelly.pocketcasts.analytics.AnalyticsTrackerWrapper
 import au.com.shiftyjelly.pocketcasts.analytics.SourceView
 import au.com.shiftyjelly.pocketcasts.discover.view.CategoryPill
 import au.com.shiftyjelly.pocketcasts.models.entity.PodcastEpisode
+import au.com.shiftyjelly.pocketcasts.models.entity.TrendingPodcast
 import au.com.shiftyjelly.pocketcasts.preferences.Settings
 import au.com.shiftyjelly.pocketcasts.repositories.playback.PlaybackManager
 import au.com.shiftyjelly.pocketcasts.repositories.podcast.EpisodeManager
@@ -38,6 +40,7 @@ import io.reactivex.rxkotlin.subscribeBy
 import io.reactivex.schedulers.Schedulers
 import java.io.InvalidObjectException
 import javax.inject.Inject
+import kotlinx.coroutines.launch
 import timber.log.Timber
 
 @HiltViewModel
@@ -118,7 +121,7 @@ class DiscoverViewModel @Inject constructor(
         loadData(resources)
     }
 
-    fun loadPodcastList(source: String): Flowable<PodcastList> {
+    fun loadPodcastList(source: String, categoryId: String): Flowable<PodcastList> {
         return repository.getListFeed(source)
             .map {
                 PodcastList(
@@ -133,15 +136,23 @@ class DiscoverViewModel @Inject constructor(
                     listId = it.listId,
                 )
             }
+            .doOnSuccess { podcastList ->
+                if (categoryId == NetworkLoadableList.TRENDING) {
+                    viewModelScope.launch {
+                        val podcastIds = podcastList.podcasts.map { TrendingPodcast(it.uuid, it.title.orEmpty()) }
+                        podcastManager.replaceTrendingPodcasts(podcastIds)
+                    }
+                }
+            }
             .flatMapPublisher { addSubscriptionStateToPodcasts(it) }
             .flatMap {
                 addPlaybackStateToList(it)
             }
     }
-    fun filterPodcasts(source: String, onPodcastsLoaded: (PodcastList) -> Unit) {
+    fun filterPodcasts(source: String, categoryId: String, onPodcastsLoaded: (PodcastList) -> Unit) {
         state.postValue(DiscoverState.FilteringPodcastsByCategory)
 
-        loadPodcastList(source).subscribeBy(
+        loadPodcastList(source, categoryId).subscribeBy(
             onNext = {
                 state.postValue(DiscoverState.PodcastsFilteredByCategory)
                 onPodcastsLoaded(it)
@@ -155,6 +166,7 @@ class DiscoverViewModel @Inject constructor(
 
     fun loadCarouselSponsoredPodcasts(
         sponsoredPodcastList: List<SponsoredPodcast>,
+        categoryId: String,
     ): Flowable<List<CarouselSponsoredPodcast>> {
         val sponsoredPodcastsSources = sponsoredPodcastList
             .filter {
@@ -167,7 +179,7 @@ class DiscoverViewModel @Inject constructor(
                 !isInvalidSponsoredSource
             }
             .map { sponsoredPodcast ->
-                loadPodcastList(sponsoredPodcast.source as String)
+                loadPodcastList(sponsoredPodcast.source as String, categoryId)
                     .filter {
                         it.podcasts.isNotEmpty() && it.listId != null
                     }