+++ modules/features/player/src/main/java/au/com/shiftyjelly/pocketcasts/player/view/PlayerHeaderFragment.kt
@@ -227,7 +227,7 @@ class PlayerHeaderFragment : BaseFragment(), PlayerClickListener {
                 )
                 (activity as? FragmentHostListener)?.let { listener ->
                     listener.closePlayer()
-                    listener.openPodcastPage(podcastUuid)
+                    listener.openPodcastPage(podcastUuid, sourceView.analyticsValue)
                 }
             }
 
@@ -363,24 +363,30 @@ class PlayerHeaderFragment : BaseFragment(), PlayerClickListener {
         return ThemeColor.playerHighlight01(viewModel.theme, viewModel.iconTintColor)
     }
 
-    private var lastLoadedBaseEpisode: BaseEpisode? = null
+    private var lastLoadedBaseEpisodeId: String? = null
     private var lastUseEpisodeArtwork: Boolean? = null
+    private var lastLoadedChapterPath: String? = null
 
     private fun loadArtwork(
         baseEpisode: BaseEpisode,
         useEpisodeArtwork: Boolean,
         imageView: ImageView,
     ) {
-        if (lastLoadedBaseEpisode == baseEpisode && lastUseEpisodeArtwork == useEpisodeArtwork) {
+        if (lastLoadedBaseEpisodeId == baseEpisode.uuid && lastUseEpisodeArtwork == useEpisodeArtwork) {
             return
         }
 
-        lastLoadedBaseEpisode = baseEpisode
+        lastLoadedBaseEpisodeId = baseEpisode.uuid
         lastUseEpisodeArtwork = useEpisodeArtwork
         imageRequestFactory.create(baseEpisode, useEpisodeArtwork).loadInto(imageView)
     }
 
     private fun loadChapterArtwork(chapter: Chapter?, imageView: ImageView) {
+        if (lastLoadedChapterPath == chapter?.imagePath) {
+            return
+        }
+
+        lastLoadedChapterPath = chapter?.imagePath
         chapter?.imagePath?.let { pathOrUrl ->
             imageRequestFactory.createForFileOrUrl(pathOrUrl).loadInto(imageView)
         } ?: run {
@@ -447,7 +453,7 @@ class PlayerHeaderFragment : BaseFragment(), PlayerClickListener {
         val podcast = viewModel.podcast
         (activity as FragmentHostListener).closePlayer()
         if (podcast != null) {
-            (activity as? FragmentHostListener)?.openPodcastPage(podcast.uuid)
+            (activity as? FragmentHostListener)?.openPodcastPage(podcast.uuid, sourceView.analyticsValue)
         } else {
             (activity as? FragmentHostListener)?.openCloudFiles()
         }