+++ modules/services/model/src/main/java/au/com/shiftyjelly/pocketcasts/models/db/AppDatabase.kt
@@ -51,6 +51,7 @@ import au.com.shiftyjelly.pocketcasts.models.entity.Podcast
 import au.com.shiftyjelly.pocketcasts.models.entity.PodcastEpisode
 import au.com.shiftyjelly.pocketcasts.models.entity.PodcastRatings
 import au.com.shiftyjelly.pocketcasts.models.entity.SearchHistoryItem
+import au.com.shiftyjelly.pocketcasts.models.entity.TrendingPodcast
 import au.com.shiftyjelly.pocketcasts.models.entity.UpNextChange
 import au.com.shiftyjelly.pocketcasts.models.entity.UpNextEpisode
 import au.com.shiftyjelly.pocketcasts.models.entity.UserEpisode
@@ -73,8 +74,9 @@ import au.com.shiftyjelly.pocketcasts.localization.R as LR
         UserEpisode::class,
         PodcastRatings::class,
         DbChapter::class,
+        TrendingPodcast::class,
     ],
-    version = 92,
+    version = 94,
     exportSchema = true,
     autoMigrations = [
         AutoMigration(from = 81, to = 82, spec = AppDatabase.Companion.DeleteSilenceRemovedMigration::class),
@@ -700,6 +702,26 @@ abstract class AppDatabase : RoomDatabase() {
             )
         }
 
+        val MIGRATION_92_93 = addMigration(92, 93) { database ->
+            database.execSQL(
+                """
+                    ALTER TABLE episode_chapters
+                    ADD COLUMN is_embedded INTEGER NOT NULL DEFAULT 0
+                """.trimIndent(),
+            )
+        }
+
+        val MIGRATION_93_94 = addMigration(93, 94) { database ->
+            database.execSQL(
+                """
+                    CREATE TABLE trending_podcasts(
+                        uuid TEXT NOT NULL PRIMARY KEY, 
+                        title TEXT NOT NULL
+                    )
+                """.trimIndent(),
+            )
+        }
+
         fun addMigrations(databaseBuilder: Builder<AppDatabase>, context: Context) {
             databaseBuilder.addMigrations(
                 addMigration(1, 2) { },
@@ -1082,6 +1104,8 @@ abstract class AppDatabase : RoomDatabase() {
                 MIGRATION_89_90,
                 MIGRATION_90_91,
                 MIGRATION_91_92,
+                MIGRATION_92_93,
+                MIGRATION_93_94,
             )
         }